RavenDB-19193 Check if we can use reactstrap with bootstrap 5 along with old bootstrap 3
