+++ src/Raven.Studio/wwwroot/Content/css/styles.less
@@ -7,8 +7,6 @@
 
 @import "./styles-common-root.less";
 
-//@import "./bootstrap/bootstrap.less";
-
 
 // Core variables and mixins
 @import "./bootstrap/mixins.less";
